+++ src/main/java/com/devexperts/uilocalizer/AddLocalizationNodesVisitor.java
@@ -4,7 +4,7 @@ package com.devexperts.uilocalizer;
  * #%L
  * UI Localizer
  * %%
- * Copyright (C) 2015 - 2016 Devexperts, LLC
+ * Copyright (C) 2015 - 2017 Devexperts, LLC
  * %%
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -55,7 +55,7 @@ class AddLocalizationNodesVisitor extends JCTree.Visitor {
             if (!localizedClasses.contains(currentClass)) {
                 localizedClasses.add(currentClass);
                 localizationNodesFactory.setPositionFor(currentClass);
-                currentClass.defs = currentClass.defs.append(localizationNodesFactory.getLocaleConstantDecl());
+                currentClass.defs = currentClass.defs.prepend(localizationNodesFactory.getLocaleConstantDecl());
                 currentClass.defs = currentClass.defs.append(localizationNodesFactory.getStringMethodDeclaration());
             }
         }