+++ src/main/java/com/devexperts/uilocalizer/Localizable.java
@@ -4,7 +4,7 @@ package com.devexperts.uilocalizer;
  * #%L
  * UI Localizer
  * %%
- * Copyright (C) 2015 - 2016 Devexperts, LLC
+ * Copyright (C) 2015 - 2017 Devexperts, LLC
  * %%
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this