+++ src/main/java/com/devexperts/uilocalizer/OutputUtil.java
@@ -4,7 +4,7 @@ package com.devexperts.uilocalizer;
  * #%L
  * UI Localizer
  * %%
- * Copyright (C) 2015 - 2016 Devexperts, LLC
+ * Copyright (C) 2015 - 2017 Devexperts, LLC
  * %%
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -39,8 +39,13 @@ public class OutputUtil {
             if (!scopesToWriters.containsKey(scope)) {
                 scopesToWriters.put(scope, new PrintWriter(scope + ".properties"));
             }
-            scopesToWriters.get(scope).println(key + "=" + e.getValue());
+            scopesToWriters.get(scope).println(key + "=" + screenSpecialSymbols(e.getValue()));
         }
         scopesToWriters.values().forEach(PrintWriter::close);
+    }
+
+    private static String screenSpecialSymbols(String value) {
+        value =  value.replaceAll("([\n=!:#])", "\\\\$1");
+        return  value.replaceAll("([\n])", "\\n");
     }
 }