+++ src/test/java/TestClass.java
@@ -2,7 +2,7 @@
  * #%L
  * UI Localizer
  * %%
- * Copyright (C) 2015 - 2016 Devexperts, LLC
+ * Copyright (C) 2015 - 2017 Devexperts, LLC
  * %%
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -15,9 +15,6 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.function.Predicate;
 
-/**
- * @author Ivan Rakov
- */
 @Deprecated
 public class TestClass {
     @Localizable("scope.key")