+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # CHANGELOG
 
+## 1.4 - 2017-05-23
+
+- FIX: Append properties to a file in natural order
+- FIX: Escape line breaks and other special symbols in a property file
+
 ## 1.3 - 2016-09-30
 
 - FIX: Static initialization code now added to every class with annotations to preserve original class loading order.